フォロー重複時にボタンUIが変わるように修正

+++ src/app/Http/Controllers/FollowsController.php
@@ -23,6 +23,13 @@ class FollowsController extends Controller
 
     public function store(Request $request)
     {
+        $follow = Follow::where('user_id', $request->userId)->where('follow_user_id', $request->followUserId)->first();
+        if($follow != null)
+        {
+            $json = self::ERROR_FOLLOWED;
+            $json['follow'] = $follow;
+            return response()->json($json, 403);
+        }
         return json_encode(['follow' => Follow::factory()->create([
             'user_id' => $request->userId,
             'follow_user_id' => $request->followUserId,
@@ -32,6 +39,10 @@ class FollowsController extends Controller
     public function destroy($id)
     {
         $follow = Follow::find($id);
+        if($follow == null)
+        {
+            return response()->json(self::ERROR_FOLLOW_IS_NOT_EXITS, 403);
+        }
         $follow->delete();
     }
 }
